fix bugs

+++ code/run.py
@@ -35,9 +35,9 @@ for n_iter in range(3):
                     print("ITER: "+str(n_iter)+","+str(n_directions)+","+str(delta)+","+str(step_size)+","+str(rlength))
                     params['delta_std'] = 0.04*delta+0.02
                     params['step_size'] = math.pow(2,step_size)*0.0001
-                    params['n_directions'] = math.pow(2,n_directions)*2
-                    params['rollout_length'] = math.pow(2,rlength)*100
-                    params['n_iter'] = math.pow(2,n_iter)*25
+                    params['n_directions'] = int(math.pow(2,n_directions)*2)
+                    params['rollout_length'] = int(math.pow(2,rlength)*100)
+                    params['n_iter'] = int(math.pow(2,n_iter)*25)
                     logz.save_params(params,"ITER: "+str(n_iter)+","+str(n_directions)+","+str(delta)+","+str(step_size)+","+str(rlength))
                     reward = run_ars(params)
                     #record
